+++ src/encoder/tests/mod.rs
@@ -62,3 +62,6 @@ mod timestamp;
 mod date;
 mod time;
 mod uuid;
+mod macaddr;
+mod macaddr8;
+