+++ src/types/implementation/mod.rs
@@ -2,3 +2,5 @@ mod timestamp;
 mod date;
 mod time;
 mod uuid;
+mod macaddr;
+mod macaddr8;