+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2]
+### Added
+- `MacAddr` type implementation for `[u8; 6]`
+- `MacAddr` type implementation for `eui48::MacAddress`
+- `MacAddr8` type implementation for `[u8; 6]`
+- `MacAddr8` type implementation for `[u8; 8]`
+
 ## [0.0.1]
 ### Added
 - First public version